adding branch and cleanup variables, and making git_path optional
If the user is running on GitHub actions, it is most likely that
a separate clone is not necessary, but instead we parse content
in the git repository base directory. If a git_path and branch are
provided, then we instead clone these paths and proceed as we did before.
Additionally, since deletion is not always required, we add a cleanup
variable that defaults to false (we dont delete the repository uless
requested by the user

Signed-off-by: vsoch <vsochat@stanford.edu>